UseC++20 (#51)
* Update CMakeLists.txt

* Update C and C++ standards in docs and CMake

Changed the development guideline to specify C17 and C++20. Updated CMakeLists.txt to set the C standard to 17 instead of 20 for consistency with the documentation.

* Remove josuttis-jthread dependency and sources

Eliminated the vendor/josuttis-jthread directory and all its files, including CMake integration and header includes. Updated build scripts and API CMakeLists to remove references to jthread, as it is no longer required.

* Update build-project.yaml

* Revert "Remove josuttis-jthread dependency and sources"

This reverts commit 57122ac9598de729d9ab0fcd58aca8d12e5cc5e0.

* Move josuttis-jthread to custom namespace for Apple

Changed the namespace of the josuttis-jthread vendor library from 'std' to 'josuttis' to avoid conflicts with libc++ on Apple platforms. Updated CMake and code to use the new namespace and conditionally include and link the library only on Apple. Adjusted usage in GoogleOAuth2Flow to use the correct namespace alias.

* Update GoogleOAuth2Flow.hpp

* Update CMakeLists.txt